[277f661a9] Fixed adapt bugs: random queues work
This changeset only addresses a couple complier warnings; remaining
description below pertains to previous changeset 277f661a9, which is
missing a description.

  - Added MsgAdapt::count_ to maintain operation order despite
    out-of-order messages
  - Added MsgCoarsen::face_level_count_ to maintain operation order
    despite out-of-order messages

Adapt

  - removed debug code frome control_adapt.cpp
  - merged reset_bounds() and initialize_self() into single initialize_bounds()
    in Adapt class
  - removed Adapt::LevelType class enum
  - added Block child face level counts to maintain operation order
  - added Adapt::count for local counter to maintain operation order
  - update face level in Adapt::set_face_level_<foo> only if counts are
    not older
  - added "min_level_" to Index is_sibling() / is_nibling() calls
  - added Adapt face_level counts for maintaining operation order

+++ src/Cello/control_adapt.cpp
@@ -572,7 +572,7 @@ void Block::adapt_recv_level()
 
 void Block::adapt_check_messages_()
 {
-  for (int i=0; i<adapt_msg_list_.size(); i++) {
+  for (size_t i=0; i<adapt_msg_list_.size(); i++) {
     MsgAdapt * msg = adapt_msg_list_[i];
     adapt_recv_level (
                       msg->adapt_step_,
@@ -604,7 +604,7 @@ void Block::adapt_recv_level
   bool changed = false;
   int level_min=0;
   performance_start_(perf_adapt_update);
-  for (int i=0; i<ofv[0].size(); i++) {
+  for (size_t i=0; i<ofv[0].size(); i++) {
 
     int if3[3] = {ofv[0][i],ofv[1][i],ofv[2][i]};
